requirements.txt
numpy >= 1.21.0
pynrrd >= 0.4.2
